Re-enable JET at Project level (#442)
Currently JET is disabled (see #437), but making a new release in this
state would be a breaking change. (JET itself is currently broken, but
that's a separate issue.) So we have to do something to put the package
back in a suitable state without breaking the ability to build docs,
etc.

This reverses most of the parts of #437 that disabled JET, though there
are still a couple of deliberately-missing pieces:
- there's a commented-out line in `workflows/ci.yml`
- `docs/make.jl` comments-out the build of the JET tutorial